+++ test/features/parsley.js
@@ -80,6 +80,18 @@ define(function () {
         expect(parsleyInstance.options.namespace).to.be('baz-');
         expect(parsleyInstance.options.foo).to.be('bar');
       });
+      it('should have a jquery API returning undefined if done on a non existing element', function () {
+        window.console.warn = sinon.spy();
+        expect($('#foo').parsley()).to.be(undefined);
+        expect(window.console.warn.called).to.be(true);
+      });
+      it('should have a jquery API that binds multiple selectors', function () {
+        $('body').append('<div id="element">'+
+          '<input type="text" id="foo" required />' +
+          '<input type="text" id="bar" required />' +
+        '</div>');
+        expect($('input').parsley().length).to.be(2);
+      });
       afterEach(function () {
         window.ParsleyConfig = { i18n: window.ParsleyConfig.i18n, validators: window.ParsleyConfig.validators };